requirements.txt
alembic==1.12.1
astroid==2.2.5
Authlib==1.2.1
cffi==1.15.1
Click==7.0
cryptography==42.0.7
ecdsa==0.13.3
Flask==1.0.2
Flask-Cors==3.0.9
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
greenlet==3.0.3
gunicorn==22.0.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.0
Mako==1.2.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==24.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.20.0
pylint==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-editor==1.0.4
python-jose==3.3.0
rsa==4.9
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
typing_extensions==4.7.1
Werkzeug==0.15.3
wrapt==1.11.1
zipp==3.15.0